refactor: make `ArtworkListNoImage` reusable
