+++ src/app/Scenes/ArtworkLists/FourUpImageLayout.tsx
@@ -1,6 +1,6 @@
 import { Box, Flex, Spacer } from "@artsy/palette-mobile"
+import { ArtworkListNoImage } from "app/Components/ArtworkLists/components/ArtworkListNoImage"
 import { ArtworkListImage } from "app/Scenes/ArtworkLists/components/ArtworkListImage"
-import { ArtworkListNoImage } from "app/Scenes/ArtworkLists/components/ArtworkListNoImage"
 import { FC } from "react"
 
 const IMAGES_PER_ROW = 2
